1, history; 2 attribute values

+++ DigitalEstate/AppDelegate.m
@@ -40,10 +40,10 @@
     
     NSUserDefaults* prefs = [NSUserDefaults standardUserDefaults];
     
-    int pass1 = [prefs integerForKey:kPassword1];
-    int pass2 = [prefs integerForKey:kPassword2];
-    int pass3 = [prefs integerForKey:kPassword3];
-    int pass4 = [prefs integerForKey:kPassword4];
+    long pass1 = [prefs integerForKey:kPassword1];
+    long pass2 = [prefs integerForKey:kPassword2];
+    long pass3 = [prefs integerForKey:kPassword3];
+    long pass4 = [prefs integerForKey:kPassword4];
     
     if (pass1 != 0 || pass2 != 0 || pass3 != 0 || pass4 != 0)
     {
